fix(ci): int64 conversion in size-cap test; enable clang SEH keywords
- hardening_test.go: MaxTotalAttachmentSize/2 arithmetic produced an int,
  makeSizedFile takes int64 (go vet).
- mapi_impl.cpp guards now require __clang__ alongside __SEH__, and
  CMakeLists.txt sets -fms-extensions for that one file — clang gates the
  __try/__except keywords behind it on windows-gnu triples even though the
  target unwinds via SEH.

+++ internal/mapi/hardening_test.go
@@ -118,7 +118,7 @@ func TestBuildFullMIME_PerFileSizeCap(t *testing.T) {
 func TestBuildFullMIME_TotalSizeCap(t *testing.T) {
 	dir := t.TempDir()
 	// Each file passes the per-file cap; together they exceed the total cap.
-	half := MaxTotalAttachmentSize/2 + 1024
+	half := int64(MaxTotalAttachmentSize)/2 + 1024
 	a := makeSizedFile(t, dir, "page1.pdf", half)
 	b := makeSizedFile(t, dir, "page2.pdf", half)
 	msg := &MailMessage{
